# apps/api/internal/infra/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
)

type Todo struct {
	ID          string
	Title       string
	Description sql.NullString
	Status      string
	DueDate     sql.NullString
	CreatedAt   string
	UpdatedAt   string
}